refactor: reorganize domain layer
