+++ src/domain/entities/AccessKeyEntity.ts
@@ -1,0 +1,8 @@
+/* c8 ignore start */
+export interface AccessKeyEntity {
+  chatId: number;
+  userId: number;
+  accessKey: string;
+  expiresAt: number;
+}
+/* c8 ignore end */
