+++ src/domain/entities/ChatConfigEntity.ts
@@ -1,0 +1,7 @@
+/* c8 ignore start */
+export interface ChatConfigEntity {
+  chatId: number;
+  historyLimit: number;
+  interestInterval: number;
+}
+/* c8 ignore end */
