+++ src/domain/entities/ChatEntity.ts
@@ -1,0 +1,6 @@
+/* c8 ignore start */
+export interface ChatEntity {
+  chatId: number;
+  title?: string | null;
+}
+/* c8 ignore end */
