+++ src/domain/entities/UserEntity.ts
@@ -1,0 +1,9 @@
+/* c8 ignore start */
+export interface UserEntity {
+  id: number;
+  username?: string | null;
+  firstName?: string | null;
+  lastName?: string | null;
+  attitude?: string | null;
+}
+/* c8 ignore end */
