+++ src/domain/repositories/SummaryRepository.interface.ts
