+++ src/repositories/sqlite/SQLiteChatRepository.ts
@@ -1,11 +1,9 @@
 import { inject, injectable } from 'inversify';
 import type { Database } from 'sqlite';
 
+import type { ChatEntity } from '../../domain/entities/ChatEntity';
+import type { ChatRepository } from '../../domain/repositories/ChatRepository.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import {
-  type ChatEntity,
-  type ChatRepository,
-} from '../interfaces/ChatRepository.interface';
 
 @injectable()
 export class SQLiteChatRepository implements ChatRepository {
