+++ src/repositories/sqlite/SQLiteChatUserRepository.ts
@@ -1,8 +1,8 @@
 import { inject, injectable } from 'inversify';
 import type { Database } from 'sqlite';
 
+import type { ChatUserRepository } from '../../domain/repositories/ChatUserRepository.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import { type ChatUserRepository } from '../interfaces/ChatUserRepository.interface';
 
 @injectable()
 export class SQLiteChatUserRepository implements ChatUserRepository {
