+++ src/repositories/sqlite/SQLiteMessageRepository.ts
@@ -1,10 +1,10 @@
 import { inject, injectable } from 'inversify';
 import type { Database } from 'sqlite';
 
+import type { MessageRepository } from '../../domain/repositories/MessageRepository.interface';
 import type { ChatMessage } from '../../services/ai/AIService.interface';
 import type { StoredMessage } from '../../services/messages/StoredMessage.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import { type MessageRepository } from '../interfaces/MessageRepository.interface';
 
 @injectable()
 export class SQLiteMessageRepository implements MessageRepository {
