+++ src/repositories/sqlite/SQLiteSummaryRepository.ts
@@ -1,8 +1,8 @@
 import { inject, injectable } from 'inversify';
 import type { Database } from 'sqlite';
 
+import type { SummaryRepository } from '../../domain/repositories/SummaryRepository.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import { type SummaryRepository } from '../interfaces/SummaryRepository.interface';
 
 @injectable()
 export class SQLiteSummaryRepository implements SummaryRepository {
