+++ src/repositories/sqlite/SQLiteUserRepository.ts
@@ -1,11 +1,9 @@
 import { inject, injectable } from 'inversify';
 import type { Database } from 'sqlite';
 
+import type { UserEntity } from '../../domain/entities/UserEntity';
+import type { UserRepository } from '../../domain/repositories/UserRepository.interface';
 import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
-import {
-  type UserEntity,
-  type UserRepository,
-} from '../interfaces/UserRepository.interface';
 
 @injectable()
 export class SQLiteUserRepository implements UserRepository {
