+++ src/services/summaries/RepositorySummaryService.ts
@@ -3,7 +3,7 @@ import { inject, injectable } from 'inversify';
 import {
   SUMMARY_REPOSITORY_ID,
   type SummaryRepository,
-} from '../../repositories/interfaces/SummaryRepository.interface';
+} from '../../domain/repositories/SummaryRepository.interface';
 import type Logger from '../logging/Logger.interface';
 import {
   LOGGER_FACTORY_ID,
