+++ test/AdminServiceImpl.test.ts
@@ -2,11 +2,11 @@ import type { Database } from 'sqlite';
 import { describe, expect, it, vi } from 'vitest';
 
 import type { DbProvider } from '../src/repositories/DbProvider';
-import type { AccessKeyRepository } from '../src/repositories/interfaces/AccessKeyRepository.interface';
-import type { ChatUserRepository } from '../src/repositories/interfaces/ChatUserRepository.interface';
-import type { MessageRepository } from '../src/repositories/interfaces/MessageRepository.interface';
-import type { SummaryRepository } from '../src/repositories/interfaces/SummaryRepository.interface';
-import type { UserRepository } from '../src/repositories/interfaces/UserRepository.interface';
+import type { AccessKeyRepository } from '../src/domain/repositories/AccessKeyRepository.interface';
+import type { ChatUserRepository } from '../src/domain/repositories/ChatUserRepository.interface';
+import type { MessageRepository } from '../src/domain/repositories/MessageRepository.interface';
+import type { SummaryRepository } from '../src/domain/repositories/SummaryRepository.interface';
+import type { UserRepository } from '../src/domain/repositories/UserRepository.interface';
 import { AdminServiceImpl } from '../src/services/admin/AdminServiceImpl';
 import type { ChatConfigService } from '../src/services/chat/ChatConfigService';
 import type { LoggerFactory } from '../src/services/logging/LoggerFactory';
